Extract folder browser file selection
